reorg and remove debug
