+++ pkg/editor/common.go
@@ -1,7 +1,6 @@
 package editor
 
 import (
-	"fmt"
 	"image/color"
 	"strings"
 
@@ -53,7 +52,6 @@ func SetRectangleColorsFromString(cp *ColorPanel, colors []string) {
 		if c != "" && c != "Off" {
 			RGBcolor, _ := common.GetRGBColorByName(c)
 			cp.Rectanges[count].FillColor = common.ConvertRGBtoNRGBA(RGBcolor)
-			fmt.Printf("Color is %s\n", c)
 			cp.Buttons[c].check.Checked = true
 		} else {
 			cp.Rectanges[count].FillColor = color.White
